Move Artwork* widgets to infodialog/widgets.py
